@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the qandidate/toggle-api package.
  *
@@ -20,19 +22,19 @@ use Symfony\Component\HttpFoundation\Response;
 
 $app = new Application();
 
-$env = new \Dotenv\Dotenv(__DIR__ . '/../');
+$env = new \Dotenv\Dotenv(__DIR__.'/../');
 $env->load();
-$env->required(array(
+$env->required([
     'TOGGLE__DEBUG',
     'TOGGLE__REDIS_DSN',
     'TOGGLE__ALLOWED_ORIGINS',
     'TOGGLE__PREFIX',
-));
+]);
 
-$app['debug']                 = getenv('TOGGLE__DEBUG');
-$app['redis_dsn']             = getenv('TOGGLE__REDIS_DSN');
+$app['debug'] = getenv('TOGGLE__DEBUG');
+$app['redis_dsn'] = getenv('TOGGLE__REDIS_DSN');
 $app['toggle.manager.prefix'] = getenv('TOGGLE__PREFIX');
-$app['allowed_origins']       = json_decode(getenv('TOGGLE__ALLOWED_ORIGINS'));
+$app['allowed_origins'] = json_decode(getenv('TOGGLE__ALLOWED_ORIGINS'));
 
 if (JSON_ERROR_NONE !== json_last_error()) {
     throw new RuntimeException('Failed to json_decode TOGGLE__ALLOWED_ORIGINS');
@@ -40,34 +42,34 @@ if (JSON_ERROR_NONE !== json_last_error()) {
 
 $app['env'] = getenv('env') ?: 'dev';
 
-if ($app['env'] === 'test') {
+if ('test' === $app['env']) {
     $app['toggle.manager.prefix'] .= '_test';
 }
 
-$app->register(new Predis\Silex\PredisServiceProvider(), array(
+$app->register(new Predis\Silex\PredisServiceProvider(), [
     'predis.parameters' => $app['redis_dsn'],
-));
+]);
 
 $app['toggle.manager.collection'] = $app->share(function ($app) {
     return new PredisCollection($app['toggle.manager.prefix'], $app['predis']);
 });
 
-$app['toggle.manager']        = $app->share(function ($app) {
+$app['toggle.manager'] = $app->share(function ($app) {
     return new ToggleManager($app['toggle.manager.collection']);
 });
 
-$app['toggle.operator_condition_serializer'] = $app->share(function($app) {
+$app['toggle.operator_condition_serializer'] = $app->share(function ($app) {
     return new OperatorConditionSerializer(new OperatorSerializer());
 });
 
-$app['toggle.serializer'] = $app->share(function($app) {
+$app['toggle.serializer'] = $app->share(function ($app) {
     return new ToggleSerializer($app['toggle.operator_condition_serializer']);
 });
 
-$app->get('/toggles', function() use ($app) {
+$app->get('/toggles', function () use ($app) {
     $toggles = $app['toggle.manager']->all();
 
-    $serializedToggles = array();
+    $serializedToggles = [];
 
     foreach ($toggles as $toggle) {
         $serializedToggles[] = $app['toggle.serializer']->serialize($toggle);
@@ -76,19 +78,19 @@ $app->get('/toggles', function() use ($app) {
     return $app->json($serializedToggles);
 });
 
-$app['request_to_toggle'] = $app->protect(function(Request $request) use ($app) {
+$app['request_to_toggle'] = $app->protect(function (Request $request) use ($app) {
     $serialized = $request->getContent();
 
     $data = json_decode($serialized, true);
 
-    if (json_last_error() != JSON_ERROR_NONE) {
+    if (JSON_ERROR_NONE != json_last_error()) {
         return null;
     }
 
     return $app['toggle.serializer']->deserialize($data);
 });
 
-$app->put('/toggles/{name}', function(Request $request, $name) use ($app) {
+$app->put('/toggles/{name}', function (Request $request, $name) use ($app) {
     if (null === $toggle = $app['request_to_toggle']($request)) {
         return new Response('Malformed json in post body.', 400);
     }
@@ -102,10 +104,10 @@ $app->put('/toggles/{name}', function(Request $request, $name) use ($app) {
     return new Response('', 204);
 });
 
-$app->delete('/toggles/{name}', function($name) use ($app) {
+$app->delete('/toggles/{name}', function ($name) use ($app) {
     $removed = $app['toggle.manager']->remove($name);
 
-    if ( ! $removed) {
+    if (!$removed) {
         return new Response(sprintf('Unable to delete toggle "%s"', $name), 400);
     }
 
