@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the qandidate/toggle-api package.
  *
@@ -17,6 +19,6 @@ class WebTestCase extends BaseWebTestCase
 {
     public function createApplication()
     {
-        return require __DIR__ . '/../../../../app/bootstrap.php';
+        return require __DIR__.'/../../../../app/bootstrap.php';
     }
 }
