@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the qandidate/toggle-api package.
  *
@@ -9,18 +11,18 @@
  * file that was distributed with this source code.
  */
 
-require_once __DIR__ . '/../vendor/autoload.php';
+require_once __DIR__.'/../vendor/autoload.php';
 
 use Asm89\Stack\Cors;
 use Symfony\Component\HttpFoundation\Request;
 
-$app = require_once __DIR__ . '/../app/bootstrap.php';
+$app = require_once __DIR__.'/../app/bootstrap.php';
 
-$stackedApp = new Cors($app, array(
+$stackedApp = new Cors($app, [
     'allowedOrigins' => $app['allowed_origins'],
-    'allowedMethods' => array('DELETE', 'GET', 'PUT', 'POST'),
-    'allowedHeaders' => array('accept', 'content-type', 'origin', 'x-requested-with'),
-));
+    'allowedMethods' => ['DELETE', 'GET', 'PUT', 'POST'],
+    'allowedHeaders' => ['accept', 'content-type', 'origin', 'x-requested-with'],
+]);
 
 $request = Request::createFromGlobals();
 
